Fix: make A* work with (y, x) coords instead of (x, y)
I don't know why that was the way it was, but we should stick to
convention dictated by our file-reading operations.
Because when reading file you first access rows (you move on Y-vertical
axis), then columns (move on X-horizontal axis) we access our maps in
the same way: map[y][x].  It's easier to lay this map down as Python
list of lists, too.

The fix includes necessary code changes to pathing algorithms, as well
as corresponding tests.